BIS-772: Fixing failing JS test.
BIS-772: Fixing failing JS test.

BIS-772: Working on different data structure, when exporting data not compatible with import. Fixing and improving tests.

BIS-772: Working on different data structure, when exporting data not compatible with import.

BIS-772: Working on different data structure, when exporting data not compatible with import. Work in progress.

BIS-772: Working on exporting a raw file, if it's the only result produced by export and the user specifies so. Added tests for this case.

BIS-772: Working on exporting a raw file, if it's the only result produced by export and the user specifies so. Fixed the tests.

BIS-772: Working on exporting a raw file, if it's the only result produced by export and the user specifies so.

BIS-772: Fixing the issue with nested zip files, when doing consecutive exports.

BIS-772: Added tests to verify that the data limit is checked.

BIS-772: Made Java tests work as intended after introduction of the data limit.

BIS-772: Working on data upload size limit.